Added support for the "start" event
Added a new Subclass of Connection called "StartEvent"
The start event is handled in a similar way to the "Update" event.
Added comments to the events. They're basically the same as the JSDoc...

Also I tried fixing the serever but it's even more broken now.
I don't know what I did.....